ENH: pull sentinel construction into a function
